requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
pydantic
python-dotenv
google-cloud-datastore
google-cloud-aiplatform
pdfplumber
google-genai
